# Conflicts:
#	src/fdtdx/__init__.py
#	src/fdtdx/core/jax/pytrees.py
#	src/fdtdx/fdtd/container.py
#	src/fdtdx/objects/detectors/mode.py
#	src/fdtdx/objects/device/parameters/continuous.py
#	src/fdtdx/objects/object.py
#	src/fdtdx/objects/static_material/polygon.py
#	src/fdtdx/objects/static_material/sphere.py

+++ src/fdtdx/objects/sources/tfsf.py
@@ -129,7 +129,7 @@ class TFSFPlaneSource(DirectionalPlaneSourceBase, ABC):
 
         center = jnp.asarray(
             [center_horizontal + horizontal_offset, center_vertical + vertical_offset],
-            dtype=jnp.float32,
+            dtype=self._config.dtype,
         ).squeeze()
         return center
